Refactor: increase visibility of `generateAndSet()`
The `generateAndSet()` method should be available to the public for use
by asynchronous processes that are handling the refresh generation of
a cached value.

An asynchronous process could generate the value and set it in
the cache directly. However, by using `generateAndSet()`, we can take
advantage of the mutex lock. Other calls to `generateAndSet()` at
about the same time will wait for the first call to complete. If the
first such nearly-simultaneous call results in the newly generated value
being set in cache, the remaining calls will be able to return the
cached value immediately without having to call the actual `generate`
callable multiple times.

This prevents a [Cache Stampede](https://en.wikipedia.org/wiki/Cache_stampede).

This is particularly beneficial for `generate` processes that take a
long time or require significant resources.