Fixed incorrect concurrent generation logic that was preventing refresh generation
The `generateAndSet()` method was only using the default `acquireLock()`
result and whether the data value was still in cache to determine if it
could return the cached value. However, the default `acquireLock()`
result would often be true, whether the data value was very recently
refreshed, was refreshed within the default mutex lock timeout
specified in the generator, but also when this was the first attempt
to refresh the value and no other process was attempting to regenerate
it. The logic in this last case allowed the existing (and still needing
to be refreshed) data value to be returned without regenerating it.

In this commit, we improve the `acquireLock()` logic and introduce a new
cache key to mark that when value was recently refreshed.

The new `acquireLock()` logic first checks to see if we can acquire a
lock immediately. If so, there are no concurrent processes attempting
to generate this data value, so we can go ahead and generate it.
Otherwise, we try to acquire a lock waiting for the other process to
finish. If we are able to acquire a lock in this manner, and the data
value is marked as recently generated, then we can return the cached
data value.

If it was not recently refreshed or does not exist in cache, then we
proceed to generate it.

Upon generating and saving the data value in cache, we also mark it
as being recently refreshed. If another process is waiting on this one
to complete, then it will be able to return the value we just cached
once we release the lock rather than having to regenerate it
immediately.