Relax instanceof checks to `self` instead of `static`
Without this change, a subclass of ReferenceProvider may not recognize
a context or invokeContext as a ReferenceProvider of another type.

The change in this commit allows subclasses of ReferenceProvider to be
used as context objects without breaking the logic that determines if an
object should be treated as a reference provider.