Add `$transientParams` to ReferenceProvider
Transient Params are method or constructor params that should be used
instead of `params` when provided. However, only `params` are meant to
be returned by `asConfigArray()` for potential serialization.

Transient parameters are useful when a method or constructor can accept
either identifiers or the objects they identify. The identifiers can be
set in `$params` for possible serialization, and the objects they
represent can be set in `$transientParams`. Then object or method
calls that occur in the same process can use the existing objects in
`$transientParams`, saving a potentially costly lookup using the
identifiers in `$params`.